Fixed builder to handle classes, structs, and data
Necessary to work with whole value objects (i.e. `Struct` and `Data`) in addition to classes. The original class-specific implementation has been delegated to the sanitizers which specialize in dealing with class, struct, and data functionality.

Reek has been updated to ignore control parameter and too many statement code smells because they are minor in nature for this implementation and necessary, in this case, to maintain readability.

Milestone: minor